fix: EventFd locking up
when the event loop was notified, EventFd drained all bytes from the
pipe. the logic went like this:

    T1: if (!active.exchange(true))
    T1:   pipe.write(1)
    T0: if (select(pipe))
    T0:   if (active.exchange(false)
    T0:     pipe.read(32)

but, once in a blue moon this led to the scenario:

    T1: if (!active.exchange(true))
    T1:   pipe.write(1)
    T0: if (select(pipe))
    T0:   if (active.exchange(false)
    T1: if (!active.exchange(true))
    T1:   pipe.write(1)
    T0:     pipe.read(32)
    active is now true with no way of getting it
    back to false, everything stops working.

read 1 byte from the pipe instead. if more data remains on the pipe, the
event loop should call callbacks again on the next iteration.

also, handle some of the failures in the write syscall.