perf: integer math for character comparisons
First of all, characters are just uint8s, unpacked to uint32s. So
conversion to float is not necessary. I don't care about sign extension
or whatever else it tries to do. Cast (a no-op) is sufficient. But more
importantly, comparing as floats is also pointless and more expensive.
Use integer math.